Set --exists-action for pip to wipe
If libraries are installed from a specific git commit, and the repo is
changed, pip will propmpt the user for input. This change sets an
action ahead of time, so the user is not prompted, and msm will not
hang. More information about the flag, and the cases to which it
applies can be found on
https://pip.pypa.io/en/stable/reference/pip/#exists-action-option.